WIP, fold_line correctness, rust bindings finished
