+++ icicle/src/fri/extern.cu
@@ -1,5 +1,9 @@
-#include "fri.cuh"
-#include "gpu-utils/device_context.cuh"
+#include "fields/field_config.cuh"
+using namespace field_config;
+
+#include "fri.cu"
+#include "utils/utils.h"
+
 
 namespace fri {
   /**
@@ -15,8 +19,8 @@ namespace fri {
    * @param ctx The device context; if the stream is not 0, then everything is run async
    * @return `cudaSuccess` if the execution was successful and an error code otherwise.
    */
-  extern "C" cudaError_t CONCAT_EXPAND(CURVE, fold_line)(extension_t* line_eval, scalar_t* domain_elements, extension_t alpha, extension_t* folded_evals, int n, FriConfig cfg) {
-    return fold_line(line_eval, domain_elements, alpha, folded_evals, n, cfg);
+  extern "C" cudaError_t CONCAT_EXPAND(FIELD, fold_line)(extension_t* line_eval, scalar_t* domain_elements, extension_t alpha, extension_t* folded_evals, int n, FriConfig& cfg) {
+    return fri::fold_line(line_eval, domain_elements, alpha, folded_evals, n, cfg);
   };
 
   /**
@@ -32,7 +36,7 @@ namespace fri {
    * @param ctx The device context; if the stream is not 0, then everything is run async
    * @return `cudaSuccess` if the execution was successful and an error code otherwise.
    */
-  extern "C" cudaError_t CONCAT_EXPAND(CURVE, fold_circle_into_line)(extension_t* circle_evals, scalar_t* domain_elements, extension_t alpha, extension_t* folded_line_evals, int n, FriConfig cfg) {
-    return fold_circle_into_line(circle_evals, domain_elements, alpha, folded_line_evals, n, cfg);
+  extern "C" cudaError_t CONCAT_EXPAND(FIELD, fold_circle_into_line)(extension_t* circle_evals, scalar_t* domain_elements, extension_t alpha, extension_t* folded_line_evals, int n, FriConfig& cfg) {
+    return fri::fold_circle_into_line(circle_evals, domain_elements, alpha, folded_line_evals, n, cfg);
   };
 } // namespace fri
